# Frontend/workload-management/src/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8088/api/v1",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/teaching-staff": {
      "get": {
        "tags": [
          "TeachingStaff"
        ],
        "operationId": "findAllTeachingStaff",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TeachingStaffResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "TeachingStaff"
        ],
        "operationId": "saveTeachingStaff",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeachingStaffRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/teaching-staff/upload": {
      "post": {
        "tags": [
          "TeachingStaff"
        ],
        "operationId": "uploadTeachingStaff",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/status-type": {
      "get": {
        "tags": [
          "StatusType"
        ],
        "operationId": "findAllStatusTypes",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StatusTypeResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "StatusType"
        ],
        "operationId": "saveStatusType",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusTypeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/semester": {
      "get": {
        "tags": [
          "semester-controller"
        ],
        "operationId": "findAllSemesters",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SemesterResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "semester-controller"
        ],
        "operationId": "saveSemester",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SemesterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/semester/{targetSemesterId}/copy-workloads/{sourceSemesterId}": {
      "post": {
        "tags": [
          "semester-controller"
        ],
        "operationId": "copyWorkloadsFromSemester",
        "parameters": [
          {
            "name": "targetSemesterId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "sourceSemesterId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/semester/{targetSemesterId}/copy-academic-ranks/{sourceSemesterId}": {
      "post": {
        "tags": [
          "semester-controller"
        ],
        "operationId": "copyAcademicRanksFromSemester",
        "parameters": [
          {
            "name": "targetSemesterId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "sourceSemesterId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
...
    "/teaching-staff/{tStaffId}": {
      "get": {
        "tags": [
          "TeachingStaff"
        ],
        "operationId": "findTeachingStaffById",
        "parameters": [
          {
            "name": "tStaffId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeachingStaffResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "TeachingStaff"
        ],
        "operationId": "deleteTeachingStaffById",
        "parameters": [
          {
            "name": "tStaffId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "TeachingStaff"
        ],
        "operationId": "updateTeachingStaffById",
        "parameters": [
          {
            "name": "tStaffId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeachingStaffRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/status-type/{statusTypeId}": {
      "get": {
        "tags": [
          "StatusType"
        ],
        "operationId": "findStatusTypeById",
        "parameters": [
          {
            "name": "statusTypeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusTypeResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "StatusType"
        ],
        "operationId": "deleteStatusTypeById",
        "parameters": [
          {
            "name": "statusTypeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "StatusType"
        ],
        "operationId": "updateStatusTypeById",
        "parameters": [
          {
            "name": "statusTypeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusTypeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/semester/{semesterId}": {
      "get": {
        "tags": [
          "semester-controller"
        ],
        "operationId": "findSemesterById",
        "parameters": [
          {
            "name": "semesterId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SemesterResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "semester-controller"
        ],
        "operationId": "deleteSemesterById",
        "parameters": [
          {
            "name": "semesterId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "semester-controller"
        ],
        "operationId": "updateSemesterById",
        "parameters": [
          {
            "name": "semesterId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SemesterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
...
    "/semester/{semesterId}/previous-year": {
      "get": {
        "tags": [
          "semester-controller"
        ],
        "operationId": "findPreviousYearSemester",
        "parameters": [
          {
            "name": "semesterId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SemesterResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RegistrationRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        },
        "required": [
          "email"
        ]
      },
      "TeachingStaffRequest": {
        "type": "object",
        "properties": {
          "authDetails": {
            "$ref": "#/components/schemas/RegistrationRequest"
          },
          "statusId": {
            "type": "integer",
            "format": "int32"
          },
          "staffFacultyId": {
            "type": "integer",
            "format": "int32"
          },
          "staffAcademicRankId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "surname": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "staffAcademicRankId",
          "staffFacultyId",
          "statusId",
          "surname"
        ]
      },
      "StatusTypeRequest": {
        "type": "object",
        "properties": {
          "statusTypeName": {
            "type": "string"
          }
        },
        "required": [
          "statusTypeName"
        ]
      },
      "SemesterRequest": {
        "type": "object",
        "properties": {
          "semesterName": {
            "type": "string",
            "enum": [
              "PAVASARIS",
              "RUDENS"
            ]
          },
          "year": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": [
          "semesterName",
          "year"
        ]
      },
...
      "AcademicRank": {
        "type": "object",
        "properties": {
          "academicRankId": {
            "type": "integer",
            "format": "int32"
          },
          "rankName": {
            "type": "string"
          },
          "abbreviation": {
            "type": "string"
          },
          "deleted": {
            "type": "boolean"
          }
        }
      },
...
      "Faculty": {
        "type": "object",
        "properties": {
          "facultyId": {
            "type": "integer",
            "format": "int32"
          },
          "facultyName": {
            "type": "string"
          },
          "facultyFullName": {
            "type": "string"
          },
          "deleted": {
            "type": "boolean"
          }
        }
      },
...
      "StatusType": {
        "type": "object",
        "properties": {
          "statusTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "statusTypeName": {
            "type": "string"
          },
          "deleted": {
            "type": "boolean"
          }
        }
      },
      "TeachingStaff": {
        "type": "object",
        "properties": {
          "teachingStaffId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "surname": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/StatusType"
          },
          "staffFaculty": {
            "$ref": "#/components/schemas/Faculty"
          },
          "staffAcademicRank": {
            "$ref": "#/components/schemas/AcademicRank"
          },
          "positionTitle": {
            "type": "string"
          },
          "staffPhoto": {
            "type": "string"
          },
          "staffFullName": {
            "type": "string"
          },
          "deleted": {
            "type": "boolean"
          },
          "rankFullName": {
            "type": "string"
          }
        }
      },
...
      "GrantedAuthority": {
        "type": "object",
        "properties": {
          "authority": {
            "type": "string"
          }
        }
      },
      "MyUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "email": {
            "type": "string"
          },
          "teachingStaff": {
            "$ref": "#/components/schemas/TeachingStaff"
          },
          "password": {
            "type": "string"
          },
          "accountLocked": {
            "type": "boolean"
          },
          "enabled": {
            "type": "boolean"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedDate": {
            "type": "string",
            "format": "date-time"
          },
          "authorities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GrantedAuthority"
            }
          },
          "name": {
            "type": "string"
          },
          "accountNonExpired": {
            "type": "boolean"
          },
          "credentialsNonExpired": {
            "type": "boolean"
          },
          "username": {
            "type": "string"
          },
          "accountNonLocked": {
            "type": "boolean"
          }
        },
        "required": [
          "email"
        ]
      },
      "TeachingStaffResponse": {
        "type": "object",
        "properties": {
          "teachingStaffId": {
            "type": "integer",
            "format": "int32"
          },
          "user": {
            "$ref": "#/components/schemas/MyUser"
          },
          "name": {
            "type": "string"
          },
          "surname": {
            "type": "string"
          },
          "staffFaculty": {
            "$ref": "#/components/schemas/Faculty"
          },
          "status": {
            "$ref": "#/components/schemas/StatusType"
          },
          "positionTitle": {
            "type": "string"
          },
          "staffAcademicRank": {
            "$ref": "#/components/schemas/AcademicRank"
          },
          "staffPhoto": {
            "type": "string",
            "format": "byte"
          },
          "staffFullName": {
            "type": "string"
          },
          "rankFullName": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "deleted": {
            "type": "boolean"
          }
        }
      },
      "StatusTypeResponse": {
        "type": "object",
        "properties": {
          "statusTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "statusTypeName": {
            "type": "string"
          },
          "deleted": {
            "type": "boolean"
          }
        }
      },
      "SemesterResponse": {
        "type": "object",
        "properties": {
          "semesterId": {
            "type": "integer",
            "format": "int32"
          },
          "semesterName": {
            "type": "string",
            "enum": [
              "PAVASARIS",
              "RUDENS"
            ]
          },
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "deleted": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}